+++ src/GitLabApiClient/IMergeRequestsClient.cs
@@ -89,7 +89,8 @@ namespace GitLabApiClient
         /// <param name="mergeRequestIid">Iid of the merge request.</param>
         /// <param name="options">MergeRequestNotes retrieval options.</param>
         /// <returns>Merge requests satisfying options.</returns>
-        Task<IList<Note>> GetNotesAsync(ProjectId projectId, int mergeRequestIid, Action<MergeRequestNotesQueryOptions> options = null);
+        Task<IList<Note>> GetNotesAsync(ProjectId projectId, long mergeRequestIid,
+            Action<MergeRequestNotesQueryOptions> options = null);
 
         /// <summary>
         /// List erge request pipelines
